requirements.txt
streamlit
pandas
openpyxl
duckdb
plotly
jinja2
python-dotenv
google-generativeai
